THRIFT-5705: Don't accumulate accumulate retry count for TSSLSocket read()
This modifies TSSLSocket::read() to make the retry count a local variable
that starts from zero in each read() call. It otherwise leaves the
counter computation logic the same. Starting from zero in each read
call matches the behavior of TSocket.

The accumulation of the retry count was a problem for use cases that
rely on read timeouts, because if the retry count hits the retry limit,
then subsequent calls will fail.

Tests:
 - Ran C++ tests with no new failures